backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pdfplumber==0.11.4
python-multipart==0.0.12
python-docx==1.1.2